refactor(common): 문자열 목록 JPA Converter 위치 이동
